Move gitmoji and logger packages internal
