+++ internal/cli/alias/delete.go
@@ -4,8 +4,8 @@ import (
 	"fmt"
 
 	"github.com/HidemaruOwO/pummit/internal/alias"
+	"github.com/HidemaruOwO/pummit/internal/logger"
 	"github.com/HidemaruOwO/pummit/internal/prompt"
-	"github.com/HidemaruOwO/pummit/pkg/logger"
 	"github.com/spf13/cobra"
 )
