+++ internal/cli/doctor.go
@@ -5,7 +5,7 @@ import (
 	"os"
 
 	"github.com/HidemaruOwO/pummit/internal/doctor"
-	"github.com/HidemaruOwO/pummit/pkg/logger"
+	"github.com/HidemaruOwO/pummit/internal/logger"
 	"github.com/fatih/color"
 	"github.com/spf13/cobra"
 )
