+++ internal/git/git.go
@@ -11,7 +11,7 @@ import (
 	"github.com/HidemaruOwO/pummit/internal/alias"
 	"github.com/HidemaruOwO/pummit/internal/config"
 	"github.com/HidemaruOwO/pummit/internal/emojis"
-	"github.com/HidemaruOwO/pummit/pkg/logger"
+	"github.com/HidemaruOwO/pummit/internal/logger"
 )
 
 type CommitMessage struct {
